Integrate the Robolectric test framework and create one test
* Fixes #5 ;-)

The tests can now be run with "mvn test", but this requires ANDROID_HOME
to be set correctly.

The tests run using Robolectric, which requires Android 21 or so.

In order to make the tests read fluently and nicely,
I upgraded to Java 7 and included AssertJ.
AssertJ allows for very nice and fluent assertion style.